+++ synccommerce/RELEASES.md
@@ -561,3 +561,13 @@ Based on:
 - [typescript v0.32.0] synccommerce
 ### Releases
 - [NPM v0.32.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.32.0 - synccommerce
+
+## 2023-08-30 00:01:17
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.74.19 (2.88.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.32.1] synccommerce
+### Releases
+- [NPM v0.32.1] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.32.1 - synccommerce